+++ coordinator/app/src/main/kotlin/linea/coordinator/app/conflation/ConflationApp.kt
@@ -398,6 +398,7 @@ class ConflationApp(
         consecutiveProvenBlobsProvider = maxBlobEndBlockNumberTracker,
         proofAggregationClient = proverClientFactory.proofAggregationProverClient(),
         metricsFacade = metricsFacade,
+        blobsRepository = blobsRepository,
       )
   }